docker/docker-compose/python-app/requirements.txt
psycopg2-binary>=2.9.2